+++ packages/amplify-graphql-auth-transformer/src/__tests__/rds-query-auth.test.ts
@@ -18,7 +18,7 @@ describe('Verify RDS Model level Auth rules on queries:', () => {
 
     const out = testTransform({
       schema: validSchema,
-      transformers: [new ModelTransformer(), new AuthTransformer({ identityPoolId: 'TEST_IDENTITY_POOL_ID' }), new PrimaryKeyTransformer()],
+      transformers: [new ModelTransformer(), new AuthTransformer(), new PrimaryKeyTransformer()],
       modelToDatasourceMap: new Map(
         Object.entries({
           Post: {
@@ -27,6 +27,9 @@ describe('Verify RDS Model level Auth rules on queries:', () => {
           },
         }),
       ),
+      synthParameters: {
+        identityPoolId: 'TEST_IDENTITY_POOL_ID',
+      },
     });
     expect(out).toBeDefined();
 
@@ -116,9 +119,12 @@ describe('Verify RDS Model level Auth rules on queries:', () => {
 
     const out = testTransform({
       schema: validSchema,
-      transformers: [new ModelTransformer(), new AuthTransformer({ identityPoolId: 'TEST_IDENTITY_POOL_ID' }), new PrimaryKeyTransformer()],
+      transformers: [new ModelTransformer(), new AuthTransformer(), new PrimaryKeyTransformer()],
       authConfig,
       modelToDatasourceMap,
+      synthParameters: {
+        identityPoolId: 'TEST_IDENTITY_POOL_ID',
+      },
     });
     expect(out).toBeDefined();
 
@@ -240,9 +246,12 @@ describe('Verify RDS Model level Auth rules on queries:', () => {
 
     const out = testTransform({
       schema: validSchema,
-      transformers: [new ModelTransformer(), new AuthTransformer({ identityPoolId: 'TEST_IDENTITY_POOL_ID' }), new PrimaryKeyTransformer()],
+      transformers: [new ModelTransformer(), new AuthTransformer(), new PrimaryKeyTransformer()],
       authConfig,
       modelToDatasourceMap,
+      synthParameters: {
+        identityPoolId: 'TEST_IDENTITY_POOL_ID',
+      },
     });
     expect(out).toBeDefined();
 
@@ -309,7 +318,7 @@ describe('Verify RDS Model level Auth rules on queries:', () => {
 
     const out = testTransform({
       schema: validSchema,
-      transformers: [new ModelTransformer(), new AuthTransformer({ identityPoolId: 'TEST_IDENTITY_POOL_ID' }), new PrimaryKeyTransformer()],
+      transformers: [new ModelTransformer(), new AuthTransformer(), new PrimaryKeyTransformer()],
       authConfig,
       modelToDatasourceMap: new Map(
         Object.entries({
@@ -319,6 +328,9 @@ describe('Verify RDS Model level Auth rules on queries:', () => {
           },
         }),
       ),
+      synthParameters: {
+        identityPoolId: 'TEST_IDENTITY_POOL_ID',
+      },
     });
     expect(out).toBeDefined();
 
@@ -370,9 +382,12 @@ describe('Verify RDS Model level Auth rules on queries:', () => {
 
     const out = testTransform({
       schema: validSchema,
-      transformers: [new ModelTransformer(), new AuthTransformer({ identityPoolId: 'TEST_IDENTITY_POOL_ID' }), new PrimaryKeyTransformer()],
+      transformers: [new ModelTransformer(), new AuthTransformer(), new PrimaryKeyTransformer()],
       authConfig,
       modelToDatasourceMap,
+      synthParameters: {
+        identityPoolId: 'TEST_IDENTITY_POOL_ID',
+      },
     });
     expect(out).toBeDefined();