+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.17.4](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.17.3...amplify-category-api-e2e-tests@3.17.4) (2023-10-21)
+
+**Note:** Version bump only for package amplify-category-api-e2e-tests
+
 ## [3.17.3](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.17.2...amplify-category-api-e2e-tests@3.17.3) (2023-10-12)
 
 **Note:** Version bump only for package amplify-category-api-e2e-tests